dom/camera/test/test_bug1099390.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bug 1099390</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="camera_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id="viewfinder" width="200" height="200" autoplay></video>
<img src="#" alt="This image is going to load" id="testimage"/>
<script class="testbody" type="text/javascript;version=1.7">

var suite = new CameraTestSuite();

suite.test('bug-1099390', function() {
  function release(p) {
    return new Promise(function(resolve, reject) {
      var gotCloseEvent = false;
      var gotReleasePromise = false;

      var onClosed = function(e) {
        suite.camera.removeEventListener('close', onClosed);
        ok(!gotCloseEvent, "gotCloseEvent was " + gotCloseEvent);
        ok(e.reason === "HardwareReleased", "'close' event reason is: " + e.reason);
        gotCloseEvent = true;
        if (gotReleasePromise) {
          resolve();
        }
      };

      suite.camera.addEventListener('close', onClosed);

      suite.camera.release().then(function(p) {
        ok(true, "released camera");
        gotReleasePromise = true;
        if (gotCloseEvent) {
          resolve();
        }
      }).catch(reject);
    });
  }

  return suite.getCamera()
    .then(release, suite.rejectGetCamera);
});

suite.setup()
  .then(suite.run);

</script>
</body>

</html>
